Use larger fonts for u-boot
The default font on MacBook is terribly small, and painful to deal with.
This changeset applies a couple of existing patches that add support for
the terminus font to u-boot, and set the config to use them.

This results in nice fonts with a very reasonable fonts on a MacBook. It
will likely result in somewhat large fonts on non-HiDPI displays, but
those should still be well readable.

I opted to patch the PKGBUILD rather than submitting this to the Asahi
u-boot branch, since this seems somewhat out-of-scope for that that
branch is focusing on, and this series would add too much noise there.

Signed-off-by: Hugo Osvaldo Barrera <hugo@barrera.io>