asahi-configs: New package
A container for mostly desktop-related configs we need to put on
systems.

Some of these are already in the images we ship, but let's make a
package so we can push new ones for upgrades.

This will be a dependency of asahi-meta, since it doesn't hurt if the
desktop stuff isn't installed and we want everyone to get these,
especially 30-modeset.conf.

Signed-off-by: Hector Martin <marcan@marcan.st>