asahi-meta: add tiny-dfr as dependency
This will install tiny-dfr and makes the touch bar work out of the box
as soon as a kernel with touch bar display and digitizer driver is
present.
The systemd service is device activated and conditional on "apple,j293"
and "apple,j493" device tree compatibles. It will not run on other
systems.

Signed-off-by: Janne Grunau <j@jannau.net>